Renamed the TreeNode enum to TreeNodeType to remove ambiguity. Moved remaning tree functions to common.
