+++ app/src/main/java/com/gatheringhallstudios/mhworlddatabase/features/weapons/detail/WeaponDetailFamilyFragment.kt
@@ -11,7 +11,7 @@ import com.gatheringhallstudios.mhworlddatabase.common.RecyclerViewFragment
 import com.gatheringhallstudios.mhworlddatabase.components.ChildDivider
 import com.gatheringhallstudios.mhworlddatabase.components.DashedDividerDrawable
 import com.gatheringhallstudios.mhworlddatabase.features.weapons.WeaponTreeListAdapterDelegate
-import com.gatheringhallstudios.mhworlddatabase.features.weapons.RenderedTreeNode
+import com.gatheringhallstudios.mhworlddatabase.common.RenderedTreeNode
 import com.gatheringhallstudios.mhworlddatabase.getRouter
 
 /**
